requirements.txt
boto3
streamlit
langchain_community
langchain_core
langchain_aws
uuid
tiktoken